// tilde/.config/karabiner/types.ts
export interface KarabinerRules {
  description?: string
  manipulators?: Manipulator[]
}

export interface Manipulator {
  description?: string
  type: 'basic'
  from: From
  to?: To[]
  to_after_key_up?: To[]
  to_if_alone?: To[]
  parameters?: Parameters
  conditions?: Conditions[]
}

export interface Parameters {
  'basic.simultaneous_threshold_milliseconds'?: number
}

type Conditions =
  | FrontMostApplicationCondition
  | DeviceCondition
  | KeybaordTypeCondition
  | InputSourceCondition
  | VaribaleCondition
  | EventChangedCondition
// ...
export interface From {
  key_code?: KeyCode
  simultaneous?: SimultaneousFrom[]
  simultaneous_options?: SimultaneousOptions
  modifiers?: Modifiers
}

export interface Modifiers {
  optional?: ModifiersKeys[]
  mandatory?: ModifiersKeys[]
}

export interface To {
  key_code?: KeyCode
  modifiers?: ModifiersKeys[]
  shell_command?: string
  set_variable?: {
    name: string
    value: boolean | number | string
  }
  mouse_key?: MouseKey
  pointing_button?: string
  /**
   * Power Management plugin
   * @example: sleep system
   * @see: {@link https://karabiner-elements.pqrs.org/docs/json/complex-modifications-manipulator-definition/to/software_function/iokit_power_management_sleep_system/}
   */
  software_function?: SoftwareFunction
}